examples/mcp-sse/pyproject.toml
[project]
name = "mcp-sse"
version = "0.1.0"
description = "SSE-based MCP Server and Client"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.2.1",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["client", "weather"]